introducing the CourseOverview component
