+++ Code/7-intro-to-react/hello-world/src/ProgressCounter.js
@@ -1,0 +1,11 @@
+export function ProgressCounter({topics}) {
+
+    const discussedTopics = topics.filter(each => each.isDiscussed);
+    const discussedCount = discussedTopics.length;
+
+    return (
+        <div>
+            Discussed: {discussedCount}
+        </div>
+    )
+}
